# MNT: Reorganize non-constant global statics into structs