… numpy/_core/src/multiarray/arrayobject.h
… numpy/_core/src/multiarray/arraytypes.c.src
… numpy/_core/src/multiarray/common.h
… numpy/_core/src/multiarray/common_dtype.c
… numpy/_core/src/multiarray/compiled_base.c
… numpy/_core/src/multiarray/conversion_utils.c
… numpy/_core/src/multiarray/convert_datatype.c
… numpy/_core/src/multiarray/convert_datatype.h
… numpy/_core/src/multiarray/ctors.c
… numpy/_core/src/multiarray/descriptor.c
… numpy/_core/src/multiarray/dlpack.c
… numpy/_core/src/multiarray/dtypemeta.c
… numpy/_core/src/multiarray/getset.c
… numpy/_core/src/multiarray/item_selection.c
… numpy/_core/src/multiarray/methods.c
… numpy/_core/src/multiarray/methods.h
… numpy/_core/src/multiarray/multiarraymodule.c
… numpy/_core/src/multiarray/multiarraymodule.h
… numpy/_core/src/multiarray/npy_static_data.c
… numpy/_core/src/multiarray/npy_static_data.h
… numpy/_core/src/multiarray/number.c
… numpy/_core/src/multiarray/scalartypes.c.src
… numpy/_core/src/multiarray/shape.c
… numpy/_core/src/multiarray/strfuncs.c
… numpy/_core/src/multiarray/stringdtype/dtype.c
… numpy/_core/src/umath/_scaled_float_dtype.c
… numpy/_core/src/umath/dispatching.c
… numpy/_core/src/umath/extobj.c
… numpy/_core/src/umath/extobj.h
… numpy/_core/src/umath/funcs.inc.src