This reorganizes most of the mutable static globals in numpy into four structs exposed to the internal API via `multiarraymodule.h`

* `npy_interned_str` for interned strings
* `npy_static_pydata` for immutable PyObjects that are initialized during module initialization
* `npy_static_cdata` for immutable C data that is initialized during module initialization
* `npy_ma_thread_unsafe_state` for state stored in a thread-unsafe manner

With the goal of refactoring the items in `npy_ma_thread_unsafe_state` to be thread-safe in followup PRs. See also the tracking issue for items that still need to be fixed.