@ngoldbaum 2024-06-04
I ran the full benchmark suite. It looks like there are some performance changes but it's kind of obnoxious to pick out reproducible changes from the big list `spin bench -c` spits out.

So far I've found:

```
| Change   | Before [a2d19725] <main>   | After [95a592b5] <global-state-refactor>   |   Ratio | Benchmark (Parameter)                                           |
|----------|----------------------------|--------------------------------------------|---------|-----------------------------------------------------------------|
| +        | 3.22±0.01μs                | 5.82±0.03μs                                |    1.81 | bench_ufunc.NDArrayAsType.time_astype(('float32', 'complex64')) |
| +        | 3.44±0.03μs                | 5.87±0.02μs                                |    1.71 | bench_ufunc.NDArrayAsType.time_astype(('float64', 'complex64')) |
```

see the full list here: https://gist.github.com/ngoldbaum/7e8ee9a129a96a32536f228e4214018b

I suspect the majority is noise. I'll try to pare it down to just the real performance changes and see if I can understand what's going on.