@ngoldbaum 2024-06-05
I tried comparing `main` with itself on the full benchmark suite using `spin bench -c` and saw no changes, so the performance changes seen above certainly aren't noise like I said earlier. Will try to dig in to understand why some of the differences end up being random large performance regressions or improvements depending on the test run...