@mattip 2024-06-05
Looks like a move in the right direction to me. I wonder if the performance changes if you statically allocate the structs rather than calling `PyMem_Calloc`. Does the dynamic allocation affect the size of the shared object or is it too small to notice?