@ngoldbaum 2024-06-05
> I wonder if the performance changes if you statically allocate the structs rather than calling PyMem_Calloc.

Good point! It does seem to help, at least on Linux. There are still some heap allocations though, I still need to look closer at whether they can be made static.