@ngoldbaum 2024-06-06
So the latest version of this PR drops moving things into the `npy_ma_state` struct to avoid heap allocations entirely. I think that does improve performance but it's difficult for me to tell which benchmarks give useful results and which don't.

Earlier I said I compared `main` with itself using `spin bench`, but I did that using `spin bench -c main` with `main` checked out, and it looks like there is some state that asv persists between runs on the same commit hash, because when I do the same thing but with two identical commits with different hashes (generated by doing `git commit --ament` with no changes to the working directory), I see random changes in the results for some benchmarks:

`` `
$ spin bench -c f3643a2d85ef4c18769608e111d465ce4a778474 -v -t "bench_linalg.Linalg"
... eliding a bunch of output ...
| Change   | Before [f3643a2d]    | After [9e40ee24] <global-state-refactor>   |   Ratio | Benchmark (Parameter)                      |
|----------|----------------------|--------------------------------------------|---------|--------------------------------------------|
| +        | 783±3μs              | 1.45±0.7ms                                 |    1.85 | bench_linalg.Linalg.time_svd('int16')      |
| +        | 32.6±0.05μs          | 45.9±10μs                                  |    1.41 | bench_linalg.Linalg.time_det('complex64')  |
| -        | 35.5±0.04μs          | 32.5±0.03μs                                |    0.92 | bench_linalg.Linalg.time_det('float32')    |
| -        | 2.18±0.06ms          | 785±0.9μs                                  |    0.36 | bench_linalg.Linalg.time_svd('float64')    |
| -        | 2.20±0.04ms          | 792±1μs                                    |    0.36 | bench_linalg.Linalg.time_svd('int64')      |
| -        | 109±70μs             | 35.9±0.2μs                                 |    0.33 | bench_linalg.Linalg.time_det('complex128') |
```

Here f3643a2d is the most recent commit in this PR and 9e40ee24 is the commit I generated with a no-op `git commit --amend`.

I'm going to start a full benchmark run like this to determine which benchmarks have random results and I will ignore those for determining whether or not this PR has a performance impact.