@ngoldbaum 2024-06-06
Nice! That makes sense. This is just one benchmark I could run quickly to prove there's an issue, I'll take a look at the full list of randomly changing tests to see if there are similar problems.