@ngoldbaum 2024-06-13
I think this is ready to merge now. Since this will definitely conflict with the `set_string_function` PR, why don't we merge that one first, I'll fix the conflicts here, and then we can merge this one?

Also if there are worries about generating lots of conflicts to the 2.0 maintenance branch I'd be happy to hold off on merging this until we're a little closer to the 2.1 maintenance branch being created. I can continue to update this as I make more things thread-safe.